+++ library/CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to the library will be documented in this file.
 - Change `_addIssue` to not ignore empty strings as error message (pull request #1065)
 - Change `ISO_DATE_TIME_REGEX` and `ISO_TIMESTAMP_REGEX` to support space as separator (pull request #1064)
 - Change pipe tuple of `pipe` and `pipeAsync` to be readonly by default
+- Change `forward`, `forwardCheck`, `partialCheck` and `partialCheckAsync` to improve TypeScript performance (issue #987)
 - Refactor `bytes`, `maxBytes`, `minBytes` and `notBytes` action
 - Fix implementation of `nonOptional`, `nonOptionalAsync`, `nonNullable`, `nonNullableAsync`, `nonNullish` and `nonNullishAsync` schema in edge cases (issue #909)
 - Fix instantiation error for `any` in `PathKeys` type (issue #929)